Add dto and change integer to long

+++ src/main/java/tinder/tindesrv/repository/PersToPersRepository.java
@@ -8,7 +8,7 @@ import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Set;
 
-public interface PersToPersRepository extends JpaRepository<PersToPers, Integer> {
+public interface PersToPersRepository extends JpaRepository<PersToPers, Long> {
     /**
      * Возвращает список любимцев найденных по user id
      *
@@ -16,7 +16,7 @@ public interface PersToPersRepository extends JpaRepository<PersToPers, Integer>
      * @return Set<Integer> id любимцев
      */
     @Query(value = "select p.crushId from PersToPers p where p.userId=?1")
-    Set<Integer> getCrushIdByUserId(Integer userId);
+    Set<Long> getCrushIdByUserId(Long userId);
 
     /**
      * Возвращает список клиентов найденных по crush id
@@ -25,7 +25,7 @@ public interface PersToPersRepository extends JpaRepository<PersToPers, Integer>
      * @return Set<Integer> id клиентов
      */
     @Query(value = "select p.userId from PersToPers p where p.crushId=?1")
-    Set<Integer> getUserIdByCrushId(Integer crushId);
+    Set<Long> getUserIdByCrushId(Long crushId);
 
     /**
      * Возвращает список клиентов с которым был мэтч
@@ -35,7 +35,7 @@ public interface PersToPersRepository extends JpaRepository<PersToPers, Integer>
      */
     @Query(value = "select p1.crushId from PersToPers p1, PersToPers p2 where p1.userId=p2.crushId " +
             "and p2.userId=p1.crushId and p1.userId=?1")
-    Set<Integer> getMatchesId(Integer userId);
+    Set<Long> getMatchesId(Long userId);
 
     /**
      * Найти всех контактов по user id или crush id
@@ -44,7 +44,7 @@ public interface PersToPersRepository extends JpaRepository<PersToPers, Integer>
      * @param crushId id любимца
      * @return List<PersToPers> лист связей
      */
-    List<PersToPers> findDistinctByUserIdOrCrushId(int userId, int crushId);
+    List<PersToPers> findDistinctByUserIdOrCrushId(Long userId, Long crushId);
 
     /**
      * Есть ли связь между клиентом и любимцем
@@ -53,7 +53,7 @@ public interface PersToPersRepository extends JpaRepository<PersToPers, Integer>
      * @param crushId id любимца
      * @return true - если есть, false - если нет
      */
-    boolean existsByUserIdAndCrushId(int userId, int crushId);
+    boolean existsByUserIdAndCrushId(Long userId, Long crushId);
 
     /**
      * Удаление связи между клиентом и любимцем
@@ -62,5 +62,5 @@ public interface PersToPersRepository extends JpaRepository<PersToPers, Integer>
      * @param crushId id любимца
      */
     @Transactional
-    void deleteByUserIdAndCrushId(Integer userId, Integer crushId);
+    void deleteByUserIdAndCrushId(Long userId, Long crushId);
 }
